// src/api/model/taskDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vector
 * The Vector API
 * OpenAPI spec version: 1.0
 */
import type { TaskDtoDeletedAt } from './taskDtoDeletedAt';
import type { TaskDtoDeletedBy } from './taskDtoDeletedBy';
import type { TaskDtoDescription } from './taskDtoDescription';
import type { TaskDtoNotes } from './taskDtoNotes';

export interface TaskDto {
  createdAt: string;
  createdBy: number;
  updatedAt: string;
  updatedBy: number;
  /** @nullable */
  deletedAt: TaskDtoDeletedAt;
  /** @nullable */
  deletedBy: TaskDtoDeletedBy;
  id: number;
  title: string;
  /** @nullable */
  description: TaskDtoDescription;
  flagged: boolean;
  deadlineType: string;
  issuedAt: string;
  dueDate: string;
  /** @nullable */
  notes: TaskDtoNotes;
  workspaceId: number;
  sourceId: number;
}